final (w/o additonal)
the jump is a problem because the update function in character isnt specific to mario so koopa also has the jump functions and things that should be specific to mario because i didnt make that class, the next thing i need ot do is make that but do that in a different branch because the code here for koopa is fundamentally okay if it wasnt also inhereiting marios functions.
once that is done, the jump should work, and do the additional work for tutorial 11 before moving onto 12.